agregado validaciones, y calculos de movimientos

+++ devsu-movimientos/src/main/java/com/pruebatecnicajava/devsu/templates/MovimientoTemp.java
@@ -25,6 +25,6 @@ public class MovimientoTemp {
     
     private String numerocuenta;
     private TipoMovimientoEnum tipoMovimiento;
-    private Long valor;
+    private BigDecimal valor;
     
 }
